# api/spec/openapi.yaml
openapi: 3.1.0
info:
  title: XHS Growth Engine API
  version: 1.0.0
  description: |
    API specification for XHS Growth Engine - a LangGraph-based multi-agent system
    for automating content growth on Xiaohongshu (小红书). Orchestrates workflow:
    trend scouting → content strategy → copywriting → visual design → human review
    → publishing → analytics → engagement.
servers:
  - url: /api
    description: API server base path
...
paths:
...
  /creator-agent/model:
    get:
      tags:
        - creator-agent
      summary: Get the current Creator Model
      description: Returns the account-scoped, revisioned creator judgement model.
      operationId: getCreatorModel
      parameters:
        - name: account_id
          in: query
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        '200':
          description: Creator Model retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_CreatorModel'
        '404':
          description: Creator Model not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
    put:
      tags:
        - creator-agent
      summary: Create or revise a Creator Model
      description: Uses optimistic concurrency; expected_revision 0 creates revision 1.
      operationId: saveCreatorModel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatorModelSaveRequest'
      responses:
        '200':
          description: Creator Model saved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_CreatorModel'
        '409':
          description: Stale Creator Model revision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
  /creator-agent/model/revisions:
    get:
      tags:
        - creator-agent
      summary: List the Creator Model Revision History
      description: >-
        Returns immutable Model Revision snapshots, newest revision first. The
        history is append-only: every Creator Model write adds exactly one
        snapshot and no write ever mutates an earlier one.
      operationId: listCreatorModelRevisions
      parameters:
        - name: account_id
          in: query
          required: true
          schema:
            type: string
            minLength: 1
        - name: cursor
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Model Revision History page retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ModelRevisionPage'
        '400':
          description: Invalid cursor or limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
  /creator-agent/model/revisions/{revision}:
    get:
      tags:
        - creator-agent
      summary: Get one immutable Creator Model revision
      description: >-
        Resolves one stored Model Revision snapshot. A missing revision never
        falls back to the current Creator Model, and foreign accounts are
        indistinguishable from absent revisions.
      operationId: getCreatorModelRevision
      parameters:
        - name: revision
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
        - name: account_id
          in: query
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        '200':
          description: Model Revision snapshot retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ModelRevision'
        '404':
          description: Creator Model revision not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
  /creator-agent/decisions/{decision_id}/model-revision:
    get:
      tags:
        - creator-agent
      summary: Resolve the Creator Model revision a decision used
      description: >-
        Returns the exact immutable Model Revision a stored Decision Record was
        evaluated against. The decision is never recomputed against the current
        model.
      operationId: getCreatorDecisionModelRevision
      parameters:
        - name: decision_id
          in: path
          required: true
          schema:
            type: string
            minLength: 1
        - name: account_id
          in: query
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        '200':
          description: Decision Model Revision resolved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ModelRevision'
        '404':
          description: Decision record or its Creator Model revision not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
  /creator-agent/dataset/decisions:
    get:
      tags:
        - creator-agent
      summary: List the Creator Agent Decision Dataset
      description: Returns immutable Decision Record snapshots with linked Learning Signal IDs.
      operationId: listCreatorDecisionDataset
      parameters:
        - name: account_id
          in: query
          required: true
          schema:
            type: string
            minLength: 1
        - name: audience_id
          in: query
          required: false
          schema:
            type: string
            minLength: 1
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/DecisionStatus'
        - name: feedback_outcome
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/FeedbackOutcome'
        - name: has_feedback
          in: query
          required: false
          schema:
            type: boolean
        - name: cursor
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Decision Dataset page retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_DecisionDatasetPage'
        '400':
          description: Invalid cursor, limit, or filter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
  /creator-agent/decisions:
    post:
      tags:
        - creator-agent
      summary: Create an evidence-backed decision
      description: Evaluates candidates with the exact Creator Model revision and persists an audit record.
      operationId: createCreatorDecision
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecisionRequest'
      responses:
        '200':
          description: Decision record created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_DecisionRecord'
        '404':
          description: Creator Model not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
  /creator-agent/decisions/{decision_id}:
    get:
      tags:
        - creator-agent
      summary: Get a persisted decision record
      operationId: getCreatorDecision
      parameters:
        - name: decision_id
          in: path
          required: true
          schema:
            type: string
        - name: account_id
          in: query
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        '200':
          description: Decision record retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_DecisionRecord'
        '404':
          description: Decision record not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
  /creator-agent/decisions/{decision_id}/feedback:
    post:
      tags:
        - creator-agent
      summary: Record feedback on a decision
      description: Appends feedback and updates relationship memory without mutating the Creator Model.
      operationId: recordCreatorFeedback
      parameters:
        - name: decision_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordFeedbackRequest'
      responses:
        '200':
          description: Feedback recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_FeedbackResult'
        '400':
          description: Feedback audience does not match the decision audience
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '404':
          description: Decision record not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
  /creator-agent/relationships/{audience_id}:
    get:
      tags:
        - creator-agent
      summary: Get relationship memory for an audience member
      operationId: getCreatorRelationship
      parameters:
        - name: audience_id
          in: path
          required: true
          schema:
            type: string
        - name: account_id
          in: query
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        '200':
          description: Relationship memory retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_RelationshipMemory'
  /creator-agent/actions:
    post:
      tags:
        - creator-agent
      summary: Plan a confirmation-gated Creator Agent action
      description: Validates and persists a non-transactional Action Intent without invoking an external system.
      operationId: planCreatorAction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActionIntentRequest'
      responses:
        '200':
          description: Action Intent created or returned from an idempotent retry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ActionIntent'
        '400':
          description: Action does not match the Decision Record or candidate set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '404':
          description: Decision record not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
    get:
      tags:
        - creator-agent
      summary: List Creator Agent Action Intents
      description: Returns account-scoped intents ordered newest first, optionally filtered by status.
      operationId: listCreatorActions
      parameters:
        - name: account_id
          in: query
          required: true
          schema:
            type: string
            minLength: 1
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ActionStatus'
      responses:
        '200':
          description: Action Intents retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ActionIntentList'
  /creator-agent/actions/{action_id}/resolve:
    post:
      tags:
        - creator-agent
      summary: Confirm or cancel a Creator Agent Action Intent
      description: Records an explicit resolution; confirmed authorizes a later explicit execution call.
      operationId: resolveCreatorAction
      parameters:
        - name: action_id
          in: path
          required: true
          schema:
            type: string
            minLength: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActionResolutionRequest'
      responses:
        '200':
          description: Action Intent resolved without external side effects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ActionIntent'
        '404':
          description: Action Intent not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        '409':
          description: Conflicting resolution for an already resolved Action Intent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
  /creator-agent/actions/{action_id}/execute:
    post:
      tags:
        - creator-agent
      summary: Execute a confirmed Creator Agent Action Intent
      description: Produces one immutable, deterministic local receipt without contacting an external system.
      operationId: executeCreatorAction
      parameters:
        - name: action_id
          in: path
          required: true
          schema:
            type: string
            minLength: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecuteActionRequest'
      responses:
        '200':
          description: Action execution receipt created or returned from an idempotent retry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ActionExecution'
        '404':
          description: Action Intent or source Decision Record not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        '409':
          description: Action Intent must be confirmed before execution
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
  /creator-agent/actions/{action_id}/execution:
    get:
      tags:
        - creator-agent
      summary: Get a Creator Agent Action execution receipt
      description: Returns an immutable receipt within the account scope.
      operationId: getCreatorActionExecution
      parameters:
        - name: action_id
          in: path
          required: true
          schema:
            type: string
            minLength: 1
        - name: account_id
          in: query
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        '200':
          description: Action execution receipt retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ActionExecution'
        '404':
          description: Execution receipt not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
...
components:
  schemas:
...
    ApiResponse_CreatorModel:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/CreatorModel'
...
    DecisionStatus:
      type: string
      enum: [recommended, insufficient_evidence, no_eligible_candidate]
...
    CreatorModelSaveRequest:
      type: object
      required: [account_id, model]
      properties:
        account_id:
          type: string
          minLength: 1
          maxLength: 128
        expected_revision:
          type: integer
          minimum: 0
          default: 0
        model:
          $ref: '#/components/schemas/CreatorModelDefinition'
...
    NotFound:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            success:
              type: boolean
              enum: [false]
            error:
              type: object
              properties:
                code:
                  type: string
                  enum: [NOT_FOUND, ERROR_CREATOR_MODEL_NOT_FOUND, ERROR_CREATOR_MODEL_REVISION_NOT_FOUND, ERROR_CREATOR_DECISION_NOT_FOUND, ERROR_CREATOR_LEARNING_SIGNAL_NOT_FOUND, ERROR_CREATOR_EVIDENCE_NOT_FOUND, ERROR_CREATOR_ACTION_NOT_FOUND, ERROR_CREATOR_ACTION_EXECUTION_NOT_FOUND]
                message:
                  type: string
          example:
            success: false
            data: null
            error:
              code: "NOT_FOUND"
              message: "Workflow not found"
              details:
                thread_id: "thread_invalid"
            timestamp: "2026-05-25T10:00:00Z"
            request_id: "req_error_002"
...